Refactor into separate Thread class
