+++ features/fixtures/macos/macOSTestApp/MainWindowController.m
@@ -93,7 +93,7 @@ static void BSLog(NSString *format, ...) NS_FORMAT_FUNCTION(1,2) NS_NO_TAIL_CALL
 
 - (IBAction)executeMazeRunnerCommand:(id)sender {
     Scenario.baseMazeAddress = @"http://localhost:9339";
-    [Scenario executeMazeRunnerCommand:^(NSString *action, NSString *scenarioName, NSString *eventMode){
+    [Scenario executeMazeRunnerCommand:^(NSString *scenarioName, NSString *eventMode){
         self.scenarioName = scenarioName;
         self.scenarioMetadata = eventMode;
     }];
